services/recs-profile/tests/pyproject.toml
[project]
name = "recs-profile-tests"
version = "0.1.0"
description = "Functional tests from recs-profile"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-asyncio>=0.26.0,<1.0.0",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-v --tb=short"
log_cli = true
log_cli_level = "INFO"
